Fix: 지원서류 조회 response 정의
