+++ src/main/java/com/gdsc_knu/official_homepage/entity/application/Application.java
@@ -57,6 +57,8 @@ public class Application extends BaseTimeEntity {
     @Enumerated(EnumType.STRING)
     private Track track;
 
+    private String note;
+
     @Builder.Default
     @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
     @JoinColumn(name = "application_id")
